Rename `circuit` module to `circuits`
